Fix an issue where the delegate method `-pubnubClient:willSuspendWithBlock:` is not called until _after_ the client resumes. This is preventing custom client work that should happen prior to the application (and PubNub) suspending.
To do this, we must move the beginBackgroundTask call outside of the async block as the main queue may be (and usually is) suspended at the end of the runloop iteration. This means the async block which should be run on the next main thread runloop iteration to not be called until the app resumes.