fix(presence): address issue with immediate `unsubscribe`
Fix the issue where an immediate 'unsubscribe' after receiving a response from a subscription may
cause a race condition where the next 'subscribe' is in the process of scheduling and can't be
cancelled, causing 'leave' to time out.

fix(cocoapods): address classes visibility

Fix the issue because of which private classes leaked into public headers and the CocoaPod build
failed.